refactor(ToDo): Move ToDoTable to dedicated directory
This change moves the `ToDoTable` and `ToDoTableRow` components into a new directory, `web/src/pages/ToDo/ToDoDesktopView/`.

The purpose of this refactoring is to improve the component's organization. Since `ToDoTable` is only used within `ToDoDesktopView`, moving it closer to its consumer makes the project structure clearer and more cohesive. It prevents the root `ToDo` directory from becoming cluttered and reinforces the logical relationship between the components.
