+++ web/src/pages/ToDo/ToDoDesktopView/ToDoTable.jsx
@@ -12,11 +12,11 @@ import { visuallyHidden } from "@mui/utils";
 import PropTypes from "prop-types";
 import { useMemo } from "react";
 
-import { useSkipUntilAuthUserIsReady } from "../../hooks/auth";
-import { useGetTicketsQuery } from "../../services/tcApi";
-import { APIError } from "../../utils/APIError";
-import { ssvcPriorityProps } from "../../utils/const";
-import { errorToString, utcStringToLocalDate } from "../../utils/func";
+import { useSkipUntilAuthUserIsReady } from "../../../hooks/auth";
+import { useGetTicketsQuery } from "../../../services/tcApi";
+import { APIError } from "../../../utils/APIError";
+import { ssvcPriorityProps } from "../../../utils/const";
+import { errorToString, utcStringToLocalDate } from "../../../utils/func";
 
 import { ToDoTableRow } from "./ToDoTableRow";
